refactor: add GoBackButton component, refactor imports
