+++ src/components/AuthenticationButtons/AuthenticationButtons.tsx
@@ -1,4 +1,4 @@
-import { Button } from '../Button';
+import { Button } from '../Buttons';
 
 import './AuthenticationButtons.scss';
