+++ src/components/CharactersList/CharactersList.tsx
@@ -1,13 +1,11 @@
-import { useNavigate } from 'react-router-dom';
-
 import { CharacterCard } from '../CharacterCard';
 import { useGetAllCharactersPerPageQuery } from '../../store';
 import { useBasicMathOperations } from '../../hooks/useBasicMathOperations';
+import { GoBackButton } from '../Buttons';
 
 import './CharactersList.scss';
 
 const CharactersList = () => {
-  const navigate = useNavigate();
   const { increment, decrement, number } = useBasicMathOperations(1, 42);
   const {
     data: characters,
@@ -18,7 +16,7 @@ const CharactersList = () => {
   if (isLoading) return <h2 className='loading-message'>Данные загружаются...</h2>;
   if (isError) return (
     <>
-      <button className='button' onClick={() => navigate(-1)}>назад</button>
+      <GoBackButton />
       <h2 className='error-message'>Что-то пошло не так. Попробуйте зайти позже!</h2>
     </>
   );
@@ -27,7 +25,12 @@ const CharactersList = () => {
     return (
       <>
         <ul className='characters-list'>
-          {characters.map(character => <CharacterCard key={character.id} character={character} />)}
+          {characters.map(character => (
+            <CharacterCard
+              key={character.id}
+              character={character}
+            />
+          ))}
         </ul>
         <div className="characters-list__button-wrapper">
           <button
