+++ src/pages/FavoritesPage/FavoritesPage.tsx
@@ -1,5 +1,4 @@
-import { useNavigate } from 'react-router-dom';
-
+import { GoBackButton } from '../../components/Buttons';
 import { CharacterCard } from '../../components/CharacterCard';
 import { useAppSelector } from '../../hooks/useAppSelector';
 import { useGetMultipleCharactersQuery } from '../../store';
@@ -7,14 +6,13 @@ import { useGetMultipleCharactersQuery } from '../../store';
 import './FavoritesPage.scss';
 
 const FavoritesPage = () => {
-  const navigate = useNavigate();
   const favorites = useAppSelector(store => store.userSlice.favorites);
   const { data: characters, isError, isLoading } = useGetMultipleCharactersQuery(favorites.toString());
 
   if (favorites.length === 0) {
     return (
       <>
-        <button className='button' onClick={() => navigate(-1)}>назад</button>
+        <GoBackButton />
         <h2 className='loading-message'>У Вас нет избранных персонажей.</h2>
       </>
     );
@@ -26,7 +24,7 @@ const FavoritesPage = () => {
   if (characters) {
     return (
       <>
-        <button className='button' onClick={() => navigate(-1)}>назад</button>
+        <GoBackButton />
         <div className='favorites-page'>
           {characters.map(character => <CharacterCard key={character.id} character={character} />)}
         </div>
