+++ src/store/index.ts
@@ -1,2 +1,3 @@
 export * from './store';
 export * from './charactersAPI';
+export * from './userSlice';
